fix(temporal): keep producer fixture dependency-free
